MVC_2_Ready

+++ src/main/java/web/Dao/UserDao.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface UserDao {
     List<User> getAllUser();
-    User show(int id);
+    User show(long id);
     void save(User user);
-    void update(int id,User user);
-    void delete(int id);
+    void update(long id,User user);
+    void delete(long id);
 }
